Alterações nos atributos das classes Edge e Graph. Crianção de método getAllNodes na classe GeoPointsDatabase

+++ app/src/main/java/com/ufpimaps/models/Edge.java
@@ -9,9 +9,9 @@ public class Edge {
     private final String id;
     private final Node source;
     private final Node destination;
-    private final int weight;
+    private final double weight;
 
-    public Edge(String id, Node source, Node destination, int weight) {
+    public Edge(String id, Node source, Node destination, double weight) {
         this.id = id;
         this.source = source;
         this.destination = destination;
@@ -28,7 +28,7 @@ public class Edge {
     public Node getSource() {
         return source;
     }
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
